requirements.txt
langchain
langchain-core
langchain-community
pypdf
pymupdf
sentence-transformers
faiss-cpu
chromadb
langchain-groq
python-dotenv
pathlib
numpy
uuid
typing
scikit-learn
fastapi
pydantic
python-multipart

